// src/app/models/product.ts
export class Product {
  _id: string;
  name: string;
  code: string;
  price: number;
  weigth: number;
  picture: string;
  kind: string;
  subCategory: string;
  active: boolean;
  storeActive: boolean;
  storeAvailability: number;
  description: string;
  promotion: number;
  createdAt: Date;
  updatedAt: Date;
}